pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "A medical chatbot application"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.0.6",
    "langchain_huggingface>=0.0.6",
    "faiss-cpu>=1.7.4",
    "pypdf>=3.0.0",
    "streamlit>=1.29.0",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.19.0",
    "torch>=2.0.0,<2.1.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.21.0"
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true
allow-prereleases = true

